+++ frontend/src/routes/reviews/reviewCard.svelte
@@ -1,5 +1,6 @@
 <script>
   import Icon from "@iconify/svelte"
+  import StarRating from "@ernane/svelte-star-rating"
 
   export let review
   export let filter
@@ -50,15 +51,27 @@
       <p class="text-sm font-bold mb-2">{building} - {place}</p>
     </div>
 
-    <div class="pb-5 px-4 text-xs text-neutral-content flex items-center justify-center gap-5">
-      <div class="rating">
-        {#each Array(Math.round(rating/2)) as _}
-          <Icon icon="ic:baseline-star" class="w-6 h-6 text-primary" />
-        {/each}
-        {#each Array(5 - Math.round(rating/2)) as _}
-          <Icon icon="ic:baseline-star" class="w-6 h-6 text-primary/20" />
-        {/each}
-      </div>
+    <div class="pb-5 px-4 text-xs flex items-center justify-center gap-5">
+      <StarRating config={{
+        readOnly: true,
+        countStars: 5,
+        range: {
+          min: 0,
+          max: 5,
+          step: 0.5
+        },
+        score: rating / 2,
+        showScore: false,
+        scoreFormat: undefined,
+        name: "",
+        starConfig: {
+          size: 24,
+          fillColor: "oklch(var(--p))",
+          strokeColor: "oklch(var(--p))",
+          unfilledColor: "oklch(var(--b3))",
+          strokeUnfilledColor: "oklch(var(--b3))"
+        }
+      }} />
       <div class="text-md font-bold">
         <span class="text-primary text-lg">{rating / 2}</span> out of 5
       </div>
@@ -94,7 +107,7 @@
 
     </div>
 
-    <div class="p-4 flex items-center text-xs text-neutral-content/70">
+    <div class="p-4 flex items-center text-xs opacity-70">
       Review by {author_name || "Anonymous"} on {formatDate(date)}
     </div>
   </div>